Move ITs to sdk-tests
