+++ sdk-tests/pom.xml
@@ -16,15 +16,15 @@
     <maven.compiler.target>17</maven.compiler.target>
     <maven.deploy.skip>true</maven.deploy.skip>
     <dapr.sdk.version>1.12.0-SNAPSHOT</dapr.sdk.version>
-    <dapr.sdk.alpha.version>0.12.0-SNAPSHOT</dapr.sdk.alpha.version>
+    <dapr.alpha.sdk.version>0.12.0-SNAPSHOT</dapr.alpha.sdk.version>
     <protobuf.output.directory>${project.build.directory}/generated-sources</protobuf.output.directory>
     <protobuf.input.directory>${project.basedir}/proto</protobuf.input.directory>
     <grpc.version>1.59.0</grpc.version>
     <protobuf.version>3.17.3</protobuf.version>
     <opentelemetry.version>1.39.0</opentelemetry.version>
     <spring-boot.version>3.3.1</spring-boot.version>
-    <logback-classic.version>1.4.12</logback-classic.version>
-    <wiremock.version>3.9.1</wiremock.version>
+    <testcontainers-dapr.version>0.12.0-SNAPSHOT</testcontainers-dapr.version>
+    <testcontainers-test.version>1.19.8</testcontainers-test.version>
   </properties>
 
   <dependencyManagement>
@@ -135,12 +135,6 @@
     </dependency>
     <dependency>
       <groupId>io.dapr</groupId>
-      <artifactId>testcontainers-dapr</artifactId>
-      <version>${dapr.sdk.alpha.version}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>io.dapr</groupId>
       <artifactId>dapr-sdk-actors</artifactId>
       <version>${dapr.sdk.version}</version>
       <scope>test</scope>
@@ -150,6 +144,34 @@
       <artifactId>dapr-sdk-springboot</artifactId>
       <version>${dapr.sdk.version}</version>
       <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.dapr.spring</groupId>
+      <artifactId>dapr-spring-core</artifactId>
+      <version>${dapr.alpha.sdk.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.dapr.spring</groupId>
+      <artifactId>dapr-spring-data</artifactId>
+      <version>${dapr.alpha.sdk.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.dapr.spring</groupId>
+      <artifactId>dapr-spring-messaging</artifactId>
+      <version>${dapr.alpha.sdk.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.dapr.spring</groupId>
+      <artifactId>dapr-spring-boot-autoconfigure</artifactId>
+      <version>${dapr.alpha.sdk.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework.data</groupId>
+      <artifactId>spring-data-keyvalue</artifactId>
     </dependency>
     <dependency>
       <groupId>org.springframework.boot</groupId>
@@ -183,6 +205,36 @@
       <groupId>org.apache.commons</groupId>
       <artifactId>commons-lang3</artifactId>
       <version>3.9</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>io.dapr</groupId>
+      <artifactId>testcontainers-dapr</artifactId>
+      <version>${testcontainers-dapr.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.testcontainers</groupId>
+      <artifactId>junit-jupiter</artifactId>
+      <version>${testcontainers-test.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.testcontainers</groupId>
+      <artifactId>postgresql</artifactId>
+      <version>${testcontainers-test.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.testcontainers</groupId>
+      <artifactId>mysql</artifactId>
+      <version>${testcontainers-test.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <!-- https://mvnrepository.com/artifact/com.mysql/mysql-connector-j -->
+    <dependency>
+      <groupId>com.mysql</groupId>
+      <artifactId>mysql-connector-j</artifactId>
+      <version>8.2.0</version>
       <scope>test</scope>
     </dependency>
     <dependency>
