+++ sdk-tests/src/test/resources/query.json
